暂存修改

+++ ali_dindin/models/partner.py
@@ -105,8 +105,9 @@ class ResPartner(models.Model):
         for res in self:
             din_userid = res.din_userid
             super(ResPartner, self).unlink()
-            self.delete_din_extcontact(din_userid)
-
+            if self.env['ir.config_parameter'].sudo().get_param('ali_dindin.din_delete_extcontact'):
+                self.delete_din_extcontact(din_userid)
+            return True
 
     @api.model
     def delete_din_extcontact(self, din_userid):
